+++ packages/database/prisma/migrations/20260926120000_admin_2fa/migration.sql
@@ -1,0 +1,3 @@
+-- AlterEnum
+ALTER TYPE "OtpPurpose" ADD VALUE 'ADMIN_2FA';
+